{{cookiecutter.repository_name}}/requirements/dev.txt
{% if cookiecutter.create_jupyter_notebook_directory|lower == "y" -%}
jupyter < 2
{% endif -%}
mike >= 2, < 3
mkdocs < 2
mkdocs-material >= 9.4, < 10
{% if cookiecutter.command_line_interface|lower == "y" -%}
mkdocs-click < 0.7
{% endif -%}
{% if cookiecutter.create_jupyter_notebook_directory|lower == "y" -%}
mkdocs-jupyter < 0.25
{% endif -%}
mkdocstrings-python < 2
{% if cookiecutter.create_jupyter_notebook_directory|lower == "y" -%}
nbmake < 2
{% endif -%}
pre-commit < 4
pytest >= 8, < 9
pytest-cov < 5
pytest-mock < 4
pytest-timeout < 3
pytest-xdist < 4